+++ Inc/audiocodecstrategy.hpp
@@ -60,50 +60,12 @@ class AudioCodecStrategy
      */
     virtual void Mute(murasaki::CodecChannel channel, bool mute = true) = 0;
     /**
-     * \brief Set the line input gain and enable the relevant mixer.
-     * \param left_gain Gain by dB. The gain value outside of the acceptable range will be saturated.
-     * \param right_gain Gain by dB. The gain value outside of the acceptable range will be saturated.
-     * \param mute set true to mute
+     *
+     * \brief send one command to CODEC
+     * \param command command data array.
+     * \details
      */
-    virtual void SetLineInputGain(float left_gain, float right_gain, bool mute = false) {
-    }
-    ;
-    /**
-     * \brief Set the aux input gain and enable the relevant mixer.
-     * \param left_gain Gain by dB. The gain value outside of the acceptable range will be saturated.
-     * \param right_gain Gain by dB. The gain value outside of the acceptable range will be saturated.
-     * \param mute set true to mute
-     */
-    virtual void SetAuxInputGain(float left_gain, float right_gain, bool mute = false) {
-    }
-    ;
-    /**
-     * \brief Set the mic input gain and enable the relevant mixer.
-     * \param left_gain Gain by dB. The gain value outside of the acceptable range will be saturated.
-     * \param right_gain Gain by dB. The gain value outside of the acceptable range will be saturated.
-     * \param mute set true to mute
-     */
-    virtual void SetMicInputGain(float left_gain, float right_gain, bool mute = false) {
-    }
-    ;
-    /**
-     * \brief Set the line output gain and enable the relevant mixer.
-     * \param left_gain Gain by dB. The gain value outside of the acceptable range will be saturated.
-     * \param right_gain Gain by dB. The gain value outside of the acceptable range will be saturated.
-     * \param mute set true to mute
-     */
-    virtual void SetLineOutputGain(float left_gain, float right_gain, bool mute = false) {
-    }
-    ;
-    /**
-     * \brief Set the headphone output gain and enable the relevant mixer.
-     * \param left_gain Gain by dB. The gain value outside of the acceptable range will be saturated.
-     * \param right_gain Gain by dB. The gain value outside of the acceptable range will be saturated.
-     * \param mute set true to mute
-     */
-    virtual void SetHpOutputGain(float left_gain, float right_gain, bool mute = false) {
-    }
-    ;
+    virtual void SendCommand(const uint8_t command[], int size) = 0;
 
  protected:
     /**