+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Note that issue #46, #56 and #77 made incompatibility from the past code.
 - [Issue 80 : The sample peripheral name of the template has camel format](https://github.com/suikan4github/murasaki/issues/80)
 - [Issue 81 : DuplexAudio::DMACallback() should use match()](https://github.com/suikan4github/murasaki/issues/81)
 - [Issue 83 : Refactoring request of template. ](https://github.com/suikan4github/murasaki/issues/83)
+- [Issue 84 : Change CODEC interface. ](https://github.com/suikan4github/murasaki/issues/84)
 
 ### Security
 ### Known Issue